Creating  a holdout dataset from the beginning
In order to avoid any kind of overoptimistic evaluations, we are
creating a holdout dataset from the beginning, different from the
validation dataset used internally in Feature Selection stage.

So this new holdout dataset is not used until the final evaluation
step.

OptiML does not receive a test-dataset any more, so it will perform
cross-validation. It makes no sense using a specific validation
dataset in the OptiML, because this dataset won't be provided by the
user, it will be sampled from the provided train-datasets